Removindo null points .

+++ src/main/java/ufrn/br/web/services/PessoaService.java
@@ -75,8 +75,17 @@ public class PessoaService {
        return pessoaRepository.findByUserNameAndPassword(pessoa.getUsername());
 
     }
-    public void remover(Pessoa pessoa) {
-        pessoaRepository.delete(pessoa);
+    public boolean remover(Model model, Pessoa pessoa) {
+        List<String> erros = new ArrayList<>();
 
+        if(!pessoa.getEmprestimos().isEmpty()){
+            erros.add(" não é possivel deletar "+ pessoa.getUsername()+" com emprestimos associados .");
+            model.addAttribute("erros", erros);
+            return false;
+        }
+        erros.add("removido com sucesso");
+        model.addAttribute("sucessos", erros);
+        pessoaRepository.delete(pessoa);
+        return false;
     }
 }
